server/Python server/requirements.txt
fastapi==0.109.2
pymongo==4.6.1
"uvicorn[standard]"==0.27.0
python-dotenv==1.0.1
motor==3.3.2
torch
transformers